+++ R/ProcessAll.R
@@ -3,7 +3,7 @@ ProcessAllRmd = function(dir =".", method = "render"){
         FileSet = list.files(dir, pattern="\\.Rmd$", full.names=TRUE) 
         }
         else stop("Specified folder does not exist.\n")
-.ProcessAll(dir = dir, method = method, FileSet)
+.ProcessAll(dir = dir, method = method, FileSet, Extension=".Rmd")
       return(invisible(NULL))
 }
 
@@ -12,21 +12,23 @@ ProcessAllMd = function(dir ="."){
         FileSet = list.files(dir, pattern="\\.md$", full.names=TRUE) 
         }
         else stop("Specified folder does not exist.\n")
-.ProcessAll(dir = dir, FileSet = FileSet)
+.ProcessAll(dir = dir, FileSet = FileSet, Extension=".md")
       return(invisible(NULL))
 }
 
-.ProcessAll = function(dir =".", method = "render", FileSet){
-      if(method=="render"){
+.ProcessAll = function(dir =".", method = "render", FileSet, Extension=".Rmd") {
         for(i in FileSet){
+Outfile = sub(Extension, ".html", i)
+if(file.mtime(i) > file.mtime(Outfile)| !file.exists(Outfile)) {
+RemoveBOM(i)
+      if(method=="render"){
           rmarkdown::render(i, output_format = "all")
           }
-        }
       else if(method=="knit2html"){
-        for(i in FileSet){
           knitr::knit2html(i)
           }
-        }
-      else stop("Specified method is not yet available.\n")
+      else {stop("Specified method is not yet available.\n")}
+}
+}
       return(invisible(NULL))
       }